Support header based auth for git operations
