+++ Sources/Directory.Packages.props
@@ -10,7 +10,7 @@
     <PackageVersion Include="Aspire.Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.0-preview.6.24214.1" />
     <PackageVersion Include="coverlet.msbuild" Version="6.0.2" />
     <PackageVersion Include="Kysect.DotnetProjectSystem" Version="0.1.16" />
-    <PackageVersion Include="Kysect.GithubUtils" Version="1.3.1" />
+    <PackageVersion Include="Kysect.GithubUtils" Version="1.3.2" />
     <PackageVersion Include="Kysect.PowerShellRunner" Version="0.2.0" />
     <PackageVersion Include="Kysect.PowerShellRunner.Configuration" Version="0.2.0" />
     <PackageVersion Include="Microsoft.AspNetCore.Components.Web" Version="8.0.4" />
