+++ Sources/Integration/Kysect.Zeya.GithubIntegration/GitCredentialType.cs
@@ -1,0 +1,7 @@
+﻿namespace Kysect.Zeya.GithubIntegration;
+
+public enum GitCredentialType
+{
+    UserPassword = 1,
+    HeaderBased = 2
+}
