+++ Sources/Integration/Kysect.Zeya.GithubIntegration/GithubIntegrationCredential.cs
@@ -5,6 +5,8 @@ namespace Kysect.Zeya.GithubIntegration;
 public class GithubIntegrationCredential
 {
     [Required]
+    public GitCredentialType AuthType { get; init; } = GitCredentialType.UserPassword;
+    [Required]
     public string GithubUsername { get; init; } = null!;
     [Required]
     public string GithubToken { get; init; } = null!;
