+++ Sources/Tests/Kysect.Zeya.Tests/Application/PolicyRepositoryValidationServiceTests.cs
@@ -45,7 +45,7 @@ public class PolicyRepositoryValidationServiceTests : IDisposable
             .AddSingleton<IRepositoryFetcher>(sp =>
             {
                 ILogger<IRepositoryFetcher> logger = sp.GetRequiredService<ILogger<IRepositoryFetcher>>();
-                var repositoryFetchOptions = new RepositoryFetchOptions("GithubUsername", "GithubToken");
+                var repositoryFetchOptions = RepositoryFetchOptions.CreateWithUserPasswordAuth("GithubUsername", "GithubToken");
                 return new RepositoryFetcher(repositoryFetchOptions, logger);
             });
