+++ Sources/Tests/Kysect.Zeya.Tests/Integration/GithubIntegration/GithubIntegrationServiceTests.cs
@@ -52,7 +52,7 @@ public class GithubIntegrationServiceTests : IDisposable
             githubIntegrationOptions.Credential,
             new GitHubClient(new ProductHeaderValue("Zeya")),
             localStoragePathFactory,
-            new RepositoryFetcher(new RepositoryFetchOptions(githubIntegrationOptions.Credential.GithubUsername, githubIntegrationOptions.Credential.GithubToken), serviceProvider.GetRequiredService<ILogger<RepositoryFetcher>>()),
+            new RepositoryFetcher(RepositoryFetchOptions.CreateWithUserPasswordAuth(githubIntegrationOptions.Credential.GithubUsername, githubIntegrationOptions.Credential.GithubToken), serviceProvider.GetRequiredService<ILogger<RepositoryFetcher>>()),
             serviceProvider.GetRequiredService<ILogger<GithubIntegrationService>>());
 
         _githubRepositoryName = new GithubRepositoryName("Kysect", "Zeya");
